+++ dataland-frontend/tests/component/utils/DataFormat.cy.ts
@@ -1,9 +1,9 @@
 import { formatBytesUserFriendly } from "@/utils/NumberConversionUtils";
 import {
   formatNumberToReadableFormat,
-  formatPercentageNumber,
+  formatPercentageNumberAsString,
   formatAmountWithCurrency,
-} from "@/utils/ValuesConversionUtils";
+} from "@/utils/Formatter.ts";
 
 describe("Unit tests for formating data", () => {
   it("Check if file size display in more readable format", () => {
@@ -22,13 +22,13 @@ describe("Unit tests for formating data", () => {
     expect(outputArray).to.deep.equal(["123,456", "654,321.12", "987,654,321", "123", "0"]);
   });
   it("Check if percentage numbers is displayed in correct format (%)", () => {
-    const sampleNumbers = [12, 65.123, 0.9876, 100, 0];
+    const sampleNumbers = [0.12, 0.65123, 12, 1, 0];
     const outputArray = sampleNumbers.map((el) => {
-      return formatPercentageNumber(el);
+      return formatPercentageNumberAsString(el);
     });
-    expect(outputArray).to.deep.equal(["12.00 %", "65.12 %", "0.99 %", "100 %", "0 %"]);
+    expect(outputArray).to.deep.equal(["12 %", "65.12 %", "1200 %", "100 %", "0 %"]);
   });
-  it("Check if amount with currency is displayed in correct format (%)", () => {
+  it("Check if amount with currency is displayed in correct format", () => {
     const sampleObjects = [
       { amount: 12, currency: "USD" },
       { amount: 132.123, currency: "CHF" },
